+++ CHANGELOG.md
@@ -6,6 +6,7 @@ ChangeLog
 -------------------
 
 - IMPROVEMENT: add a "session-end" hook for removing the session storage again
+- IMPROVEMENT: store tasks per project and not per user
 
 0.0.27 (2026-05-16)
 -------------------